Added log entries for each app start/stop, handling for duped gunicorn servers on start
Each application now writes to a general log upon each start/stop. The
application manager also writes a warning when SIGINT is received, and
the HTTP application writes an ERROR upon the gunicorn webserver already
running when attempting to start. Added handling to the start sequence
of the HTTP application which checks if a gunicorn server is running, and
if it is does not let HUME start, but writes the PID(s) that need to be
killed to hume.log.